+++ spring-boot-api/src/main/java/com/ssafy/springbootapi/domain/user/api/UserController.java
@@ -5,7 +5,6 @@ import com.ssafy.springbootapi.domain.user.dto.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -15,18 +14,18 @@ public class UserController {
     private final UserService userService;
 
     @PostMapping
-    public ResponseEntity<UserSignUpResponseDTO> signUp(@RequestBody UserSignUpRequestDTO userSignUpRequestDTO){
-        return ResponseEntity.status(HttpStatus.CREATED).body(userService.signUp(userSignUpRequestDTO));
+    public ResponseEntity<UserSignUpResponse> signUp(@RequestBody UserSignUpRequest userSignUpRequest){
+        return ResponseEntity.status(HttpStatus.CREATED).body(userService.signUp(userSignUpRequest));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<UserInfoResponseDTO> getInfo(@PathVariable Long id){
+    public ResponseEntity<UserInfoResponse> getInfo(@PathVariable Long id){
         return ResponseEntity.status(HttpStatus.ACCEPTED).body(userService.getUserInfo(id));
     }
 
     @PutMapping
-    public ResponseEntity<UserUpdateResponseDTO> update(@RequestBody UserUpdateRequestDTO userUpdateRequestDTO){
-        return ResponseEntity.status(HttpStatus.OK).body(userService.updateUserInfo(userUpdateRequestDTO));
+    public ResponseEntity<UserUpdateResponse> update(@RequestBody UserUpdateRequest userUpdateRequest){
+        return ResponseEntity.status(HttpStatus.OK).body(userService.updateUserInfo(userUpdateRequest));
     }
 
     @DeleteMapping("/{id}")
