+++ spring-boot-api/src/main/java/com/ssafy/springbootapi/domain/user/dto/UserSignUpResponse.java
@@ -1,6 +1,5 @@
 package com.ssafy.springbootapi.domain.user.dto;
 
-import com.ssafy.springbootapi.domain.user.domain.User;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -8,7 +7,7 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 @Builder
-public class UserSignUpResponseDTO {
+public class UserSignUpResponse {
     private String email;
     private String name;
 }
