+++ spring-boot-api/src/main/java/com/ssafy/springbootapi/domain/user/dto/UserUpdateResponse.java
@@ -1,17 +1,13 @@
 package com.ssafy.springbootapi.domain.user.dto;
 
-import com.ssafy.springbootapi.domain.user.domain.Address;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.util.List;
-
 @AllArgsConstructor
 @Getter
 @Builder
-public class UserInfoResponseDTO {
+public class UserUpdateResponse {
     private String email;
     private String name;
-    private List<Address> addresses;
 }
